% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{scalar}{number b}
}
\value{
the greatest common divisor of a and b
}
\description{
`euclidean()` Calculate the greatest common divisor between numbers a and b.
}
\examples{
euclidean(a,b)

}
\seealso{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
