% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab05.R
\name{get_data}
\alias{get_data}
\title{Use kolada API}
\usage{
get_data(
  link = "http://api.kolada.se/v2/data/",
  kpi = "",
  municipality = "",
  period = ""
)
}
\arguments{
\item{link}{A string}

\item{kpi}{A string}

\item{municipality}{A string}

\item{period}{A string}
}
\description{
Returns the data of Kolada database, where kpi, municipality, period are parameters,2 of them should be given
}
\seealso{
https://github.com/Hypergene/kolada
}
