% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a dataframe}

\item{W}{the weight the knapsack can take,a interger}
}
\value{
a list contain the best value and specific plan(how to choose the items)
}
\description{
`knapsack_dynamic()`solve the knapsack problem by dynamic programming
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
  data.frame(
    w=sample(1:4000, size = n, replace = TRUE),
    v=runif(n = n, 0, 10000)
  )

knapsack_dynamic(x = knapsack_objects[1:500,], W = 3500)

}
\seealso{
https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem
}
