\name{lab04-package}
\alias{lab04-package}
\alias{lab04}
\docType{package}
\title{
\packageTitle{lab04}
}
\description{
\packageDescription{lab04}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab04}
\packageIndices{lab04}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab04}

Maintainer: \packageMaintainer{lab04}
}

