% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{the formula}

\item{data}{dataframe}
}
\value{
a list including Regressions coefficients,The fitted values,The residuals,The degrees of freedom,The residual variance,The variance of the regression coefficients,The t-values for each coefficient and The p-values for each coefficient
}
\description{
`linreg()`  return an object with of class linreg.
}
