% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{linreg}
\usage{
\method{plot}{linreg}(x)
}
\arguments{
\item{x}{the list}
}
\value{
plotList
}
\description{
`linreg()`  return an object with of class linreg.
}
