% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{linreg}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{x}{the list}
}
\value{
coefficients
}
\description{
`linreg()`  return an object with of class linreg.
}
