% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{linreg}
\usage{
\method{resid}{linreg}(x)
}
\arguments{
\item{x}{the list}
}
\value{
residuals
}
\description{
`linreg()`  return an object with of class linreg.
}
