% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_df.R
\name{return_df}
\alias{return_df}
\title{return_df}
\usage{
return_df(listl)
}
\arguments{
\item{listl}{the data of Kolada database}
}
\value{
The dataframe data
}
\description{
\code{return_df}  return the dataframe data
}
