% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_json.R
\name{return_json}
\alias{return_json}
\title{return_json}
\usage{
return_json(listl)
}
\arguments{
\item{listl}{the data of Kolada database}
}
\value{
the content of response
}
\description{
\code{return_json()}  return the content of response
}
