% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{linreg}
\usage{
\method{summary}{linreg}(x)
}
\arguments{
\item{x}{the list}
}
\value{
summary
}
\description{
`linreg()`  return an object with of class linreg.
}
