% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{vertices and paths of a graph}
\format{
A data frame contain 3 vectors,v1,v2 are pairs of vertices,and(v1,v2) is the edges.w are paths of the edges.  
\describe{
  \item{v1}{vertices}
  \item{v2}{vertices}
  \item{w}{paths of the edges}
  ...
}
}
\source{
\url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\usage{
wiki_graph
}
\description{
A dataset containing the vertices and paths of a graph named wiki_graph
}
\keyword{datasets}
